Update factor sources list to add Ledger devices
