+++ RadixWallet/Features/SecurityCenterFeature/SecurityFactors/FactorSourcesList+View.swift
@@ -171,6 +171,8 @@ private extension View {
 			.enterMnemonic(with: destinationStore)
 			.addMnemonic(with: destinationStore)
 			.changeMain(with: destinationStore)
+            .addNewP2PLinkSheet(with: destinationStore)
+            .noP2PLinkAlert(with: destinationStore)
 	}
 
 	private func detail(with destinationStore: PresentationStoreOf<FactorSourcesList.Destination>) -> some View {
@@ -202,4 +204,14 @@ private extension View {
 			ChangeMainFactorSource.View(store: $0)
 		}
 	}
+    
+    private func addNewP2PLinkSheet(with destinationStore: PresentationStoreOf<FactorSourcesList.Destination>) -> some View {
+        sheet(store: destinationStore.scope(state: \.addNewP2PLink, action: \.addNewP2PLink)) {
+            NewConnection.View(store: $0)
+        }
+    }
+    
+    private func noP2PLinkAlert(with destinationStore: PresentationStoreOf<FactorSourcesList.Destination>) -> some View {
+        alert(store: destinationStore.scope(state: \.noP2PLink, action: \.noP2PLink))
+    }
 }
