+++ RadixWallet/Features/SecurityCenterFeature/SecurityFactors/FactorSourcesList.swift
@@ -8,6 +8,9 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 		let kind: FactorSourceKind
 		var rows: [Row] = []
 		var selected: Row?
+        
+        var hasAConnectorExtension: Bool = false
+        var pendingAction: ActionRequiringP2P? = nil
 
 		init(context: Context = .display, kind: FactorSourceKind) {
 			self.context = context
@@ -23,6 +26,11 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 
 	typealias Action = FeatureAction<Self>
 
+    enum ActionRequiringP2P: Sendable, Hashable {
+        case addLedger
+        case continueWithFactorsource(FactorSource)
+    }
+    
 	enum ViewAction: Sendable, Equatable {
 		case task
 		case rowTapped(State.Row)
@@ -35,6 +43,7 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 	enum InternalAction: Sendable, Equatable {
 		case setSecurityProblems([SecurityProblem])
 		case setEntities([EntitiesLinkedToFactorSource])
+        case hasAConnectorExtension(Bool)
 	}
 
 	enum DelegateAction: Sendable, Equatable {
@@ -49,6 +58,9 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 			case enterMnemonic(ImportMnemonicsFlowCoordinator.State)
 			case addMnemonic(ImportMnemonic.State)
 			case changeMain(ChangeMainFactorSource.State)
+            case noP2PLink(AlertState<NoP2PLinkAlert>)
+            case addNewP2PLink(NewConnection.State)
+            case addNewLedger(AddLedgerFactorSource.State)
 		}
 
 		@CasePathable
@@ -58,6 +70,9 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 			case enterMnemonic(ImportMnemonicsFlowCoordinator.Action)
 			case addMnemonic(ImportMnemonic.Action)
 			case changeMain(ChangeMainFactorSource.Action)
+            case noP2PLink(NoP2PLinkAlert)
+            case addNewP2PLink(NewConnection.Action)
+            case addNewLedger(AddLedgerFactorSource.Action)
 		}
 
 		var body: some ReducerOf<Self> {
@@ -76,12 +91,20 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 			Scope(state: \.changeMain, action: \.changeMain) {
 				ChangeMainFactorSource()
 			}
+            Scope(state: \.addNewP2PLink, action: \.addNewP2PLink) {
+                NewConnection()
+            }
+            Scope(state: \.addNewLedger, action: \.addNewLedger) {
+                AddLedgerFactorSource()
+            }
 		}
 	}
 
 	@Dependency(\.securityCenterClient) var securityCenterClient
 	@Dependency(\.factorSourcesClient) var factorSourcesClient
+    @Dependency(\.radixConnectClient) var radixConnectClient
 	@Dependency(\.errorQueue) var errorQueue
+    @Dependency(\.ledgerHardwareWalletClient) var ledgerHardwareWalletClient
 
 	var body: some ReducerOf<Self> {
 		Reduce(core)
@@ -95,8 +118,12 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 	func reduce(into state: inout State, viewAction: ViewAction) -> Effect<Action> {
 		switch viewAction {
 		case .task:
-			return securityProblemsEffect()
-				.merge(with: entitiesEffect(state: state))
+             var effects = securityProblemsEffect()
+                .merge(with: entitiesEffect(state: state))
+            if state.kind == .ledgerHqHardwareWallet {
+                effects = effects.merge(with: checkP2PLinkEffect())
+            }
+            return effects
 
 		case let .rowTapped(row):
 			switch state.context {
@@ -142,7 +169,9 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 						wordCount: .twentyFour
 					)
 				)
-			case .ledgerHqHardwareWallet, .offDeviceMnemonic, .arculusCard, .password:
+            case .ledgerHqHardwareWallet:
+                return performActionRequiringP2PEffect(.addLedger, in: &state)
+            case .offDeviceMnemonic, .arculusCard, .password:
 				// NOTE: Added `.device` support as placeholder, but not adding the logic for ledger (which we already support)
 				// since Matt mentioned we will probably always present this screen: https://zpl.io/wyqB6Bd
 				// and I don't want to add all the logic for checking if there is a CE or not just to migrate it later.
@@ -152,6 +181,9 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 			return .none
 
 		case let .continueButtonTapped(factorSource):
+            if state.kind == .ledgerHqHardwareWallet {
+                return performActionRequiringP2PEffect(.continueWithFactorsource(factorSource), in: &state)
+            }
 			return .send(.delegate(.selectedFactorSource(factorSource)))
 
 		case .changeMainButtonTapped:
@@ -172,7 +204,10 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 			state.entities = entities
 			setRows(state: &state)
 			return .none
-		}
+        case let .hasAConnectorExtension(hasCE):
+            state.hasAConnectorExtension = hasCE
+            return .none
+        }
 	}
 
 	func reduce(into state: inout State, presentedAction: Destination.Action) -> Effect<Action> {
@@ -187,11 +222,69 @@ struct FactorSourcesList: Sendable, FeatureReducer {
 		case .changeMain(.delegate(.updated)):
 			state.destination = nil
 			return entitiesEffect(state: state)
+            
+        case let .noP2PLink(alertAction):
+            switch alertAction {
+            case .addNewP2PLinkTapped:
+                state.destination = .addNewP2PLink(.init())
+                return .none
+
+            case .cancelTapped:
+                return .none
+            }
+
+        case let .addNewP2PLink(.delegate(newP2PAction)):
+            switch newP2PAction {
+            case let .newConnection(connectedClient):
+                state.destination = nil
+                return .run { _ in
+                    try await radixConnectClient.updateOrAddP2PLink(connectedClient)
+                } catch: { error, _ in
+                    loggerGlobal.error("Failed P2PLink, error \(error)")
+                    errorQueue.schedule(error)
+                }
+            }
 
 		default:
 			return .none
 		}
 	}
+    
+    private func checkP2PLinkEffect() -> Effect<Action> {
+        .run { send in
+            for try await isConnected in await ledgerHardwareWalletClient.isConnectedToAnyConnectorExtension() {
+                guard !Task.isCancelled else { return }
+                await send(.internal(.hasAConnectorExtension(isConnected)))
+            }
+        } catch: { error, _ in
+            loggerGlobal.error("failed to get links updates, error: \(error)")
+        }
+    }
+
+    private func performActionRequiringP2PEffect(_ action: ActionRequiringP2P, in state: inout State) -> Effect<Action> {
+        // If we don't have a connection, we remember what we were trying to do and then ask if they want to link one
+        guard state.hasAConnectorExtension else {
+            state.pendingAction = action
+            state.destination = .noP2PLink(.noP2Plink)
+            return .none
+        }
+
+        state.pendingAction = nil
+
+        // If we have a connection, we can proceed directly
+        switch action {
+        case .addLedger:
+            state.destination = .addNewLedger(.init())
+            return .none
+        case let .continueWithFactorsource(fs):
+            return .send(.delegate(.selectedFactorSource(fs)))
+//            if case let .accountRecovery(olympia: olympia) = state.context {
+//                return .send(.delegate(.choseLedgerForRecovery(ledger, isOlympia: olympia)))
+//            } else {
+//                return .send(.delegate(.choseLedger(ledger)))
+//            }
+        }
+    }
 }
 
 private extension FactorSourcesList {
